Avoid PyType_Name in CYTHON_COMPILING_IN_LIMITED_API
In the limited API `PyTypeObject` is an opaque struct so we cannot access the `tp_name` field of a type, and the `PyType_Name` API isn't available either.

This changes the code to use `PyObjectGetAttrString(obj, "__name__")` in combination with the %V modifier in `PyErr_Format` to format the error messages when `CYTHON_COMPILING_IN_LIMITED_API` is enabled.

Based on Facebook D20202965