Do not modify traceback in CYTHON_COMPILING_IN_LIMITED_API
The limited API does not expose the `_PyTraceback_Add` function or any other function that allows to add additional entries to a stacktrace. Also headers other than `Python.h` are not part of PEP384 so we cannot include "framobject.h" or "traceback.h".

This change disables stacktrace modification when compiling in the limited API.

Based on Facebook D20225672