Add __Pyx_PyCode_New stub when CYTHON_COMPILING_IN_LIMITED_API
Cython injects pseudo code objects in some places so some introspection like `function.__code__.co_argcount` works.

The limited API does not provide a `PyCode_New` function and the related code object constants. This adds a stub implementation for this case that just returns None.

This is a first version of this, eventually we should write code that creates an object that behaves similar to a code object and provides the `co_xxx` attributes as well.

Based on Facebook D20309302