Consider builtin types as opaque when importing in limited API
cython supports a construct like this to declare members of object types:
```
ctypedef class __builtin__.py_long [object PyLongObject]:
    cdef digit* ob_digit
```
Among other things the generated code grabs the type and checks that tp_basicsize matches the size of the struct specified in the declaration to make sure there are no discrepancies between the struct definition used and the binary.

All of this does not play nice with the limited API in which (except for PyObject*) all builtin type structs are opaque types. This disables the size checking for types from `builtins`/`__builtins__` when `CYTHON_COMPILING_IN_LIMITED_API` is enabled. This will work fine for many files that do not use the type at all after all or are just using pointers to the type but not accessing members.

If code happens to access members then they will get compile errors from the C compiler because the struct types are actually incomplete. That's no worse than before and I think we just have to forbid using things like for PyRo as I cannot think of a sane way to enable this in the limited API.

Based on Facebook D20344150