Avoid using _PyLong_FromByteArray when CYTHON_COMPILING_IN_LIMITED_API
The limited API does not have the `_PyLong_FromByteArray` call. This part of the code is only used when oversized integers such as `__int128_t` gcc/clang extensions are passed to/from cython. Just disable this code for now.

This also introduces a `static_assert` construct to cython so we can report a compile error instead of just aborting at runtime in case of oversized integers.

Based on Facebook D20308851