Introduce ScopeWithCleanupNode
This introduces the `ScopeWithCleanupNode` utility node. It is intended to be used to ensure that some cleanup action like freeing memory is always executed.

It has two attributes `body` and `cleanup`. It will generate code that generates the `body` code followed by the `cleanup` code, it also generates the `cleanup` code on the error code path.

Based on Facebook D20374848