Fix and simplify __Pyx_FetchCommonType()
It turns out I broke and unnecessarily complicated `__Pyx_FetchCommonType` in D20202965. This was not noticable because all the features depending on it do not work with PEP384 yet anyway.

Simplify the code: I realized that PEP384 code will only ever need `__Pyx_FetchCommonTypeFromSpec` and non-PEP384 will only ever need `__Pyx_FetchCommonType` and we do not need to support in-between variants.

Based on Facebook D21644736