Generalize code managing extension types
Moving types from global variables into the module state requires the usual bookkeeping of implementing modules traverse and clear functions and providing defines.

This was already done manually for `CyFunction` and `FusedFunction` but I need to do the same for `BoundMethod` in this stack and in the future we will also need it for `AsyncGen`, `Coroutine`, `Generator` and `StopAsyncIteration`. These are enough examples that I decided to abstract the code and only maintain 1 list of common extension types used by cython.

Based on Facebook D21756553