Stop including structmember.h in newer CPython versions
See also: https://bugs.python.org/issue2897

`structmember.h` is not officially part of the stable ABI so we should
not include it when `CYTHON_COMPILING_IN_LIMITED_API`. However we have
little choice to avoid this in older API versions. And starting with
CPython 3.10 `structmember.h` is deprecated.

Based on Facebook D21756717