Fix global assignment for non-modules
The `NameNode` node performs assignments for global things. It is used for modules but also for members of types.

When adapting for the limited API the code was changed to output:
```
#if CYTHON_COMPILING_IN_LIMITED_API
PyModule_AddObject(...)
#else
...
#endif
```
which is correct for module globals but does not make sense for type members. Change the code to only do the `PyModule_AddObject` when assigning module globals.

Based on Facebook D21663625