Avoid using tp_dictoffset check in LIMITED API
Skip the shortcut that tested whether instances had a dictionary and types are immutable in the limited API: All (cython) types are created via `PyType_FromSpec` in that case and are therefore mutable anyway.

This avoids a problem where the `tp_dictoffset` field of the opaque `PyTypeObject` type was accessed.

Based on Facebook D21863698