Keep 'module dict' around for assignment to function __globals__
PyRo technically has no "module dict" available, but `PyModule_GetDict` returns a module proxy which is still useful for some things. For example we can use it to initialize the `__globals__` fields of `CythonFunction`s.

It also simplifies the code by eliminating special cases for the limited API when `Naming.moddict_cname` can be assumed to be available.

Based on Facebook D21905582