Remove self field from CythonFunction
It turns out the `m_self` field of the embedded `PyCFunction` in `CythonFunction` is always set to the function itself. So in the `CYTHON_LIMITED_API` version without `PyCFunction` we can just as well remove the field and not populate it.

This fixes a problem we currently have where the self field is not increfed before we assign it but incrementing it, leads to cycles. Anyway seems we can just remove it instead.

Based on Facebook D21911305