Stop storing types in the module state twice
For each added type the limited API code would add an entry named `entry.type.typeobj_cname` and another for `entry.type.typeptr_cname` to the module state. The first one is the name of the `PyTypeObject` for the non-limited code paths, it seems pointless to store a `PyTypeObject*` in it in the limited API when we also have `typeptr_cname` around. This change stops adding the `typeobj_cname` when `CYTHON_COMPILING_IN_LIMITED_API`.

This simplifies the code and fixes a reference count problem where we incremented the type only once but stored it in two fields and `Py_VISIT`ed it twice.

Based on Facebook D21936712