Fix setting of __dictoffset__ in the Limited API
When creating `cdef class` types for the Limited API we cannot set `tp_dictoffset`. CPython 3.9 provides a workaround for this by setting the `__dictoffset__` member in the types `tp_members` list.

This changes renamed `generate_slot_specs` to `find_slots` and change it be a pre-processing step that creates a `slots` dictionary. The code the post-processes this `slots` dictionary to pull out `dictoffset` and `flags` and handle them in a custom way. Code generation is now performed after this.

Remove `tp_getattro` special casing from the Limited API code: The non-limited code would set the slot to `__Pyx_PyObject_GenericGetAttr` to increase performance in cases where no `__dict__` was set. The limited API was always setting it, when it was missing, which is not compatible with `__dict__` usage. As this is just a performance optimization I decided to skip it for now.

Based on Facebook D21939959